+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.8.1/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.8.1/Steam.Game.Idler_1.8.1_x64-setup.exe)
 
 ## Changelog
-- #109: Fix an issue where SteamUtility *(idle)* processes were not being closed when exiting SGI
-- Improve theme styles
+- Migrate app data dir to Tauri v2 format